+++ common/src/main/java/com/zigythebird/playeranim/animation/PlayerAnimManager.java
@@ -27,9 +27,9 @@ public class PlayerAnimManager extends AnimationStack {
 	private Map<DataTicket<?>, Object> extraData;
 	private final AbstractClientPlayer player;
 
-	private double lastUpdateTime;
+	private float lastUpdateTime;
 	private boolean isFirstTick = true;
-	private double firstTickTime = -1;
+	private float firstTickTime = -1;
 	private float tickDelta;
 
 	public PlayerAnimManager(AbstractClientPlayer player) {
@@ -48,19 +48,19 @@ public class PlayerAnimManager extends AnimationStack {
 		getBoneSnapshotCollection().clear();
 	}
 
-	public double getLastUpdateTime() {
+	public float getLastUpdateTime() {
 		return this.lastUpdateTime;
 	}
 
-	public void updatedAt(double updateTime) {
+	public void updatedAt(float updateTime) {
 		this.lastUpdateTime = updateTime;
 	}
 
-	public double getFirstTickTime() {
+	public float getFirstTickTime() {
 		return this.firstTickTime;
 	}
 
-	public void startedAt(double time) {
+	public void startedAt(float time) {
 		this.firstTickTime = time;
 	}
